Upgrade to Dynamic Approach
Instead of relying on hardcoding each Lord's units, we now use a more dynamic approach where we go over every applicable Lord and attempt to map their units to Provincial versions.

If the unit does not have a mapping, then we will replace it with the same unit.

We do this to replace all units in the Lord's army, resulting in the same army composition (no matter what it was previously) but with Provincial infantry.